Add second level resolution
Right now, we just say "less than a minute ago" because we were only
getting minute level resolution from the server. But can request second
level with the `tmres` query paramter. So this updates the client to ask
for second resolution and shows this to the user. It gives them a little
more context on whether they are looking at live data or some kind of
ghost vehicle. Instead of showing the exact number of seconds I'm
rounding the number of seconds to the nearest 5 so we don't overexercise
react.